FireFox Fix
I think should ***finally*** kill the problem with video playback on FireFox. There still seem to be *certain* videos that it doesn't like. I think this has something to do with the source vids that I used and converted over into webm. However, for the most part, all of the videos should be able to play correctly with the small fix I put in to re-start their playback if they are found to be in a state that shows that it has ended on the same clock cycle that it was scheduled to play on.

Fingers Crossed!