More Animation Updates
Further optimized the css to stop animation playback when they are hidden (for the most part). there are few animations that might still play for a tick, but most of the bad ones are now fully stopped and will be disabled as one might expect when they are hidden. It is my hope that this further improves mobile performance.